Implement knowledge base loading and chunking utilities, along with RAG pipeline for experiments E1 to E4
- Added `kb_loader.py` for chunking Markdown documents into structured chunks.
- Introduced `models.py` to define data models for queries and document chunks.
- Created pipeline modules for each experiment (E1 to E4) with respective logic for processing queries and managing embeddings.
- Implemented a `Reranker` class for improving retrieval results in E3 and E4 experiments.
- Developed a `VectorStore` class for managing vector database operations using Qdrant.
- Updated `main.py` to integrate the new pipeline structure.
- Enhanced metrics calculation in `metrics_calculator.py` for better precision.
- Revised documentation to reflect changes in the experiment structure and metrics.

+++ 4_evaluation/evaluation_lib/metrics_calculator.py
@@ -148,9 +148,9 @@ def calculate_latency_metrics_from_experiments(
     llm_times = [r.llm_time_ms for r in experiment_results]
     total_times = [r.total_time_ms for r in experiment_results]
 
-    mean_retrieval = np.mean(retrieval_times)
-    mean_llm = np.mean(llm_times)
-    mean_total = np.mean(total_times)
+    mean_retrieval = float(np.mean(retrieval_times))
+    mean_llm = float(np.mean(llm_times))
+    mean_total = float(np.mean(total_times))
 
     return mean_retrieval, mean_llm, mean_total
